[CPDLP-1688] Rename delivery partners service queries class
